Hard-coding known good values for short-term
Until we can definitively figure out what size image data DeepAR is okay with, I've hardcoded a 4:3 ratio for iPad, and 1280x720 for any other device.
This is challenging as not all devices are 16:9, but I haven't had any success using the phone's width/height to resize the images.
We may need to finally reach out to DeepAR and ask to see their code for loading images into memory.