PAIGE-270: Fixed the image sizing before and after DeepAR
Turns out that the issue was that the images, once loaded, needed to be aligned relative to a 'bytes per pixel' value of 16 (4 channels * 4 bytes per channel).
After resizing the image to match this requirement, it can safely be saved out and reloaded without any 'zooming' issues (which was related to a mismatch of input and output sizes.