PAIGE-303: Updated for uncompressed data (Android)
To increase the likelihood of face detection, we need to send the actual raw image bytes to DeepAR without any additional compression.
Previously we would compress the images back as a PNG or JPG and then load them into DeepAR via their path, but this extra step of compression and saving could damage the image such that DeepAR would have difficulties detecting faces.